requirements.txt
Django>=5.2,<5.3
psycopg2-binary>=2.9
python-dotenv>=1.0
django-cors-headers>=4.3,<5
djangorestframework>=3.15,<3.16
Pillow>=10.3,<11